requirements.txt
opencv-python==4.8.0.74
mediapipe==0.10.32
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
nltk==3.8.1
spacy==3.6.0
transformers==4.30.2
pyttsx3==2.90
SpeechRecognition==3.10.0
pyspellchecker==0.7.2
language-tool-python==2.7.1
gradio==3.41.0
flask==2.3.2
pillow==10.0.0
imageio==2.31.1
moviepy==1.0.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0